Gave BCC a slider-appropriate level of precision; moved age/trigCast to InterfaceFunctions
