+++ ParameterTabs/npiParams.py
@@ -4,14 +4,12 @@
 
 # Imports
 import logging
-from typing import Literal, cast
 
-import numpy as np
 import pandas as pd
 import streamlit as st
 from pydantic import ValidationError
 
-from ClientResources.InterfaceFunctions import paramError
+from ClientResources.InterfaceFunctions import paramError, trigCast
 from ClientResources.ModelSchema import (
     Parameters,
     ageScenarioParameters,
@@ -1107,6 +1105,7 @@ the day on which BCC will return to normal.
             min_value=0.0,
             max_value=8.0,
             value=0.2,
+            step=0.05,
             disabled=not useBCCToggle,
             on_change=saveKey,
             args=["bccReducedRate", id],
@@ -1376,75 +1375,6 @@ Total") or in each individual school (for
                         """,
                         True,
                     )
-
-
-# TODO: Move to InterfaceFunctions if needed here too
-def ageCast(x: str) -> Literal[
-    "young_infant",
-    "infant",
-    "young_child",
-    "child",
-    "adolescent",
-    "young_adult",
-    "adult",
-    "older_adult",
-    "senior",
-    "older_senior",
-]:
-    """
-    Simple function to cast age strings into literals for validation.
-
-    Parameters:
-        x (str): The string to be cast.
-
-    Returns:
-        Literal: A literal with the same value as the string.
-    """
-    return cast(
-        Literal[
-            "young_infant",
-            "infant",
-            "young_child",
-            "child",
-            "adolescent",
-            "young_adult",
-            "adult",
-            "older_adult",
-            "senior",
-            "older_senior",
-        ],
-        x,
-    )
-
-
-def trigCast(x: str) -> Literal[
-    "none",
-    "timed",
-    "per_school_cases",
-    "community_cases",
-    "community_rate",
-    "per_primary_high_school_cases",
-]:
-    """
-    Simple function to cast trigger strings into literals for validation.
-
-    Parameters:
-        x (str): The string to be cast.
-
-    Returns:
-        Literal: A literal with the same value as the string.
-    """
-    return cast(
-        Literal[
-            "none",
-            "timed",
-            "per_school_cases",
-            "community_cases",
-            "community_rate",
-            "per_primary_high_school_cases",
-        ],
-        triggerConditions[x],
-    )
 
 
 def npiSaveSchema(schema: Parameters, id: int = 0, advanced: bool = False):
